moved utilities into separate files, created useToggle
