+++ orion-ui/tests/utilities/index.ts
@@ -1,3 +1,16 @@
 export * from './fixtures'
+export * from './locate'
 export * from './pages'
-export * from './utilities'
+export * from './useButton'
+export * from './useCheckbox'
+export * from './useCombobox'
+export * from './useContextSidebar'
+export * from './useForm'
+export * from './useIconButtonMenu'
+export * from './useLabel'
+export * from './useLink'
+export * from './useModal'
+export * from './usePageHeading'
+export * from './useSelect'
+export * from './useTable'
+export * from './useTag'
